# bridge-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ChainGuard Bridge API
  description: |
    REST API bridge for interacting with ChainGuard CRE workflows.
    
    ## Request Types
    
    ### Normal HTTP Requests
    Standard HTTP requests that interact with the CRE workflow in simulation mode.
    These requests follow typical REST patterns and are suitable for:
    - Monitoring contract status
    - Managing contract configurations
    - Triggering workflow simulations
    
    ### Confidential HTTP Requests (Experimental)
    CRE confidential requests execute in secure enclaves and support:
    - Secret injection via templates: `{{.secretName}}`
    - Response encryption with AES-256-GCM
    - Simulation-only (not available in deployed workflows)
    
    Note: Confidential endpoints are prefixed with `/cre/confidential/`
    
  version: 1.0.0
  contact:
    name: ChainGuard Team
    email: support@chainguard.example.com
  license:
    name: MIT

servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.chainguard.example.com
    description: Production server

tags:
  - name: health
    description: Health check endpoints
  - name: contracts
    description: Smart contract monitoring and management
  - name: alerts
    description: Alert management and history
  - name: scan
    description: On-demand contract scanning
  - name: cre-normal
    description: CRE Normal HTTP triggers (standard requests)
  - name: cre-confidential
    description: CRE Confidential HTTP triggers (experimental, simulation only)

paths:
  /health:
    get:
      tags:
        - health
      summary: Health check
      description: Check if the API server is running
      operationId: getHealth
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              example:
                status: "ok"
                timestamp: "2024-01-15T10:30:00Z"
                version: "1.0.0"

  /api/contracts:
    get:
      tags:
        - contracts
      summary: List monitored contracts
      description: Retrieve all contracts currently being monitored by ChainGuard
      operationId: listContracts
      responses:
        '200':
          description: List of monitored contracts
          content:
            application/json:
              schema:
                type: object
                properties:
                  contracts:
                    type: array
                    items:
                      $ref: '#/components/schemas/MonitoredContract'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - contracts
      summary: Add new contract to monitor
      description: Register a new smart contract for monitoring
      operationId: addContract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddContractRequest'
      responses:
        '201':
          description: Contract added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/contracts/{address}:
    get:
      tags:
        - contracts
      summary: Get contract details
      description: Retrieve detailed information about a specific monitored contract
      operationId: getContract
      parameters:
        - $ref: '#/components/parameters/ContractAddress'
      responses:
        '200':
          description: Contract details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoredContract'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - contracts
      summary: Update contract configuration
      description: Update risk thresholds or monitoring parameters for a contract
      operationId: updateContract
      parameters:
        - $ref: '#/components/parameters/ContractAddress'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContractRequest'
      responses:
        '200':
          description: Contract updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - contracts
      summary: Remove contract from monitoring
      description: Stop monitoring a specific smart contract
      operationId: deleteContract
      parameters:
        - $ref: '#/components/parameters/ContractAddress'
      responses:
        '200':
          description: Contract removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/contracts/{address}/status:
    get:
      tags:
        - contracts
      summary: Get contract risk status
      description: Get the current risk assessment and monitoring status for a contract
      operationId: getContractStatus
      parameters:
        - $ref: '#/components/parameters/ContractAddress'
      responses:
        '200':
          description: Contract status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractStatus'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/alerts:
    get:
      tags:
        - alerts
      summary: Get alert history
      description: Retrieve historical alerts with optional filtering
      operationId: getAlerts
      parameters:
        - name: address
          in: query
          description: Filter by contract address
          schema:
            type: string
            pattern: '^0x[a-fA-F0-9]{40}$'
        - name: severity
          in: query
          description: Filter by alert severity
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, CRITICAL]
        - name: limit
          in: query
          description: Maximum number of alerts to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of alerts to skip (pagination)
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  alerts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
                  total:
                    type: integer
                    description: Total number of alerts matching filters
                  limit:
                    type: integer
                  offset:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /api/scan:
    post:
      tags:
        - scan
      summary: Trigger on-demand scan
      description: |
        Execute an immediate risk scan for specified contracts.
        This is a **normal HTTP request** that triggers the CRE workflow in simulation mode.
      operationId: triggerScan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanRequest'
      responses:
        '202':
          description: Scan initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /cre/trigger:
    post:
      tags:
        - cre-normal
      summary: CRE HTTP Trigger (Normal)
      description: |
        Standard CRE HTTP trigger endpoint for workflow execution.
        
        **CRE Request Type: Normal HTTP**
        
        This endpoint simulates how CRE workflows receive HTTP triggers in production.
        In deployed workflows, this would require:
        - Authorization with EVM signature
        - Authorized keys configured in workflow
        
        For simulation, authorization is optional.
      operationId: creHttpTrigger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CREHttpTriggerPayload'
      responses:
        '200':
          description: Workflow triggered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CRETriggerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - evmSignature: []

  /cre/confidential/trigger:
    post:
      tags:
        - cre-confidential
      summary: CRE Confidential HTTP Trigger (Experimental)
      description: |
        Confidential CRE HTTP trigger for secure enclave execution.
        
        **CRE Request Type: Confidential HTTP (Experimental)**
        
        This endpoint simulates CRE confidential requests that:
        - Execute in secure enclaves (TEE)
        - Support secret injection via templates: `{{.secretName}}`
        - Encrypt responses with AES-256-GCM
        - **Simulation-only** - not available in deployed workflows
        
        Example secret injection in headers:
        ```json
        {
          "headers": {
            "Authorization": "Bearer {{.GEMINI_API_KEY}}"
          }
        }
        ```
      operationId: creConfidentialTrigger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CREConfidentialTriggerPayload'
      responses:
        '200':
          description: Confidential workflow triggered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CREConfidentialResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '501':
          description: Not available in deployed workflows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Confidential HTTP is experimental and only available in simulation mode"
      security:
        - evmSignature: []

  /cre/report:
    post:
      tags:
        - cre-normal
      summary: Submit CRE Report
      description: |
        Endpoint for CRE workflows to submit monitoring reports.
        
        **CRE Request Type: Normal HTTP POST**
        
        Supports multiple report submission patterns:
        1. Raw body with signature
        2. Concatenated signatures in body
        3. Signatures in headers
        4. JSON-formatted report
        
        Uses `cacheSettings` to prevent duplicate submissions:
        ```typescript
        cacheSettings: {
          readFromCache: true,
          maxAgeMs: 60000 // 1 minute
        }
        ```
      operationId: submitReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CREReportSubmission'
      responses:
        '201':
          description: Report submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Duplicate report (rejected by cache)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          enum: [ok, degraded, error]
        timestamp:
          type: string
          format: date-time
        version:
          type: string
        uptime:
          type: number
          description: Server uptime in seconds

    MonitoredContract:
      type: object
      required:
        - address
        - name
        - protocol
        - riskThresholds
      properties:
        address:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
          example: "0x1234567890123456789012345678901234567890"
        name:
          type: string
          example: "Aave V3 Pool"
        protocol:
          type: string
          example: "Aave"
        chain:
          type: string
          default: "ethereum"
          example: "ethereum"
        riskThresholds:
          $ref: '#/components/schemas/RiskThresholds'
        priceFeeds:
          type: array
          items:
            $ref: '#/components/schemas/PriceFeed'
        alertChannels:
          type: array
          items:
            type: string
            enum: [email, slack, webhook]
        addedAt:
          type: string
          format: date-time

    RiskThresholds:
      type: object
      required:
        - volatility
        - liquidity
        - concentration
      properties:
        volatility:
          type: number
          format: double
          minimum: 0
          maximum: 1
          example: 0.15
        liquidity:
          type: number
          format: double
          minimum: 0
          maximum: 1
          example: 0.20
        concentration:
          type: number
          format: double
          minimum: 0
          maximum: 1
          example: 0.25
        overall:
          type: number
          format: double
          minimum: 0
          maximum: 1
          example: 0.30

    PriceFeed:
      type: object
      required:
        - asset
        - feedAddress
      properties:
        asset:
          type: string
          example: "ETH/USD"
        feedAddress:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
          example: "0x5f4eC3Df9cbd43714FE2740f5E3616155c5b8419"
        decimals:
          type: integer
          default: 8

    AddContractRequest:
      type: object
      required:
        - address
        - name
        - protocol
      properties:
        address:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
        name:
          type: string
        protocol:
          type: string
        chain:
          type: string
          default: "ethereum"
        riskThresholds:
          $ref: '#/components/schemas/RiskThresholds'
        priceFeeds:
          type: array
          items:
            $ref: '#/components/schemas/PriceFeed'
        alertChannels:
          type: array
          items:
            type: string
            enum: [email, slack, webhook]

    UpdateContractRequest:
      type: object
      properties:
        name:
          type: string
        riskThresholds:
          $ref: '#/components/schemas/RiskThresholds'
        priceFeeds:
          type: array
          items:
            $ref: '#/components/schemas/PriceFeed'
        alertChannels:
          type: array
          items:
            type: string

    ContractStatus:
      type: object
      required:
        - address
        - riskLevel
        - lastChecked
      properties:
        address:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
        riskLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        riskScore:
          type: number
          format: double
          minimum: 0
          maximum: 1
        lastChecked:
          type: string
          format: date-time
        metrics:
          $ref: '#/components/schemas/RiskMetrics'
        activeAlerts:
          type: integer
          minimum: 0

    RiskMetrics:
      type: object
      properties:
        volatility:
          type: number
          format: double
        liquidity:
          type: number
          format: double
        concentration:
          type: number
          format: double
        tvl:
          type: number
          format: double
          description: Total Value Locked in USD

    Alert:
      type: object
      required:
        - id
        - contractAddress
        - severity
        - message
        - timestamp
      properties:
        id:
          type: string
          format: uuid
        contractAddress:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        resolved:
          type: boolean
          default: false
        resolvedAt:
          type: string
          format: date-time

    ScanRequest:
      type: object
      properties:
        addresses:
          type: array
          items:
            type: string
            pattern: '^0x[a-fA-F0-9]{40}$'
          description: Specific contracts to scan (empty = all)
        priority:
          type: string
          enum: [low, normal, high]
          default: normal

    ScanResponse:
      type: object
      required:
        - scanId
        - status
      properties:
        scanId:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed]
        estimatedCompletion:
          type: string
          format: date-time
        contractsQueued:
          type: integer

    CREHttpTriggerPayload:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [scan, monitor, evaluate]
          description: Workflow action to execute
        contractAddress:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
        parameters:
          type: object
          additionalProperties: true
        signature:
          type: string
          description: EVM signature for authorization (required in production)

    CREConfidentialTriggerPayload:
      type: object
      required:
        - action
        - enclaveConfig
      properties:
        action:
          type: string
          enum: [analyze, evaluate]
        contractAddress:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
        enclaveConfig:
          type: object
          description: Configuration for confidential execution
          properties:
            secretsRequired:
              type: array
              items:
                type: string
              example: ["GEMINI_API_KEY", "SLACK_WEBHOOK"]
            encryptResponse:
              type: boolean
              default: false
        parameters:
          type: object
          additionalProperties: true

    CRETriggerResponse:
      type: object
      required:
        - workflowId
        - status
      properties:
        workflowId:
          type: string
          format: uuid
        status:
          type: string
          enum: [accepted, running, completed]
        result:
          type: object
          additionalProperties: true

    CREConfidentialResponse:
      type: object
      required:
        - workflowId
        - status
        - encrypted
      properties:
        workflowId:
          type: string
          format: uuid
        status:
          type: string
          enum: [accepted, running, completed]
        encrypted:
          type: boolean
          description: Whether response is encrypted
        encryptedData:
          type: string
          description: Base64-encoded encrypted response (if encrypted=true)
        result:
          type: object
          additionalProperties: true
          description: Plaintext result (if encrypted=false)

    CREReportSubmission:
      type: object
      required:
        - reportType
        - data
      properties:
        reportType:
          type: string
          enum: [risk_assessment, alert, status_update]
        data:
          type: object
          additionalProperties: true
        signatures:
          type: array
          items:
            type: string
          description: DON node signatures (for consensus verification)
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp

    ApiSuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
          additionalProperties: true

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details:
          type: string
        code:
          type: string

  parameters:
    ContractAddress:
      name: address
      in: path
      required: true
      description: Ethereum contract address
      schema:
        type: string
        pattern: '^0x[a-fA-F0-9]{40}$'
      example: "0x1234567890123456789012345678901234567890"

  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Invalid contract address format"
            code: "VALIDATION_ERROR"

    Unauthorized:
      description: Missing or invalid authorization
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Missing EVM signature"
            code: "UNAUTHORIZED"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Contract not found"
            code: "NOT_FOUND"

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"

  securitySchemes:
    evmSignature:
      type: apiKey
      in: header
      name: X-EVM-Signature
      description: |
        EVM signature for request authorization.
        
        In production CRE workflows, HTTP triggers require:
        1. Authorized keys configured in workflow (ECDSA_EVM type)
        2. Request signed with authorized private key
        3. Signature included in header
        
        For simulation, this is optional.